requirements.txt
# Core Framework
streamlit==1.31.0

# Database & Auth
supabase>=2.22.0

# AI/LLM
groq==0.4.2

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.4

# Document Generation
reportlab==4.0.9
python-docx==1.1.0
pillow==10.2.0

# Web Scraping (LinkedIn)
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
playwright==1.41.0

# Validation & Data
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# HTML Sanitization
bleach==6.1.0

# HTTP & Utilities
requests==2.31.0
python-dotenv==1.0.0
